refactor: corrected bitbucket event type export location
