+++ packages/deployment-service/src/bitbucket/index.ts
@@ -1,2 +1,3 @@
 export * from './bitbucket-provider'
 export * from './module'
+export * from './bitbucket-event'
